requirements.txt
asgiref==3.9.2
click==8.1.8
Django==4.2.24
django-filter==25.1
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
psycopg2-binary==2.9.10
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
gunicorn